+++ Makefile
@@ -1,11 +1,11 @@
 CC=gcc
 CFLAGS=-g -Wall -Werror -O3 
-OBJS=src/main.o src/matrix.o src/ui.o 
+OBJS=src/main.o
 
 all: main
 
 main: $(OBJS)
-	$(CC) $(CFLAGS) -o $@ $^ -lncurses
+	$(CC) $(CFLAGS) -o $@ $^ -lncurses -lm
 
 %.o: %.c
 	$(CC) $(CFLAGS) -c $^ -o $@
